+++ oneflow/core/common/id_util.h
@@ -18,6 +18,7 @@ limitations under the License.
 
 #include "oneflow/core/common/util.h"
 #include "oneflow/core/common/device_type.pb.h"
+#include "oneflow/core/job/task.pb.h"
 #include <limits>
 
 namespace oneflow {
