+++ oneflow/core/device/cuda_stream_index.cpp
@@ -1,0 +1,7 @@
+#include "oneflow/core/device/cuda_stream_index.h"
+
+namespace oneflow {
+
+REGISTER_STREAM_INDEX_GENERATOR(DeviceType::kGPU, CudaStreamIndexGenerator);
+
+}  // namespace oneflow
